refactor: group Code Text files
