+++ packages/sdk-components-react/src/basic.template.tsx
@@ -55,17 +55,6 @@ export const Checkbox: TemplateMeta = {
   template: <$.Checkbox aria-label="Enable weekly summary" />,
 };
 
-export const CodeText: TemplateMeta = {
-  category: "typography",
-  template: (
-    <$.CodeText
-      code={'const status = "ready";'}
-      language="javascript"
-      theme="github-light"
-    />
-  ),
-};
-
 export const Fragment: TemplateMeta = {
   category: "general",
   template: (
