+++ packages/sdk-components-react/src/build-hooks.ts
@@ -1,3 +1,3 @@
-import { codeTextBuildHook } from "./code-text.build";
+import { codeTextBuildHook } from "./code-text/build";
 
 export const buildHooks = [codeTextBuildHook];
