+++ packages/sdk-components-react/src/code-text/options.test.ts
@@ -1,7 +1,7 @@
 import { expect, test } from "vitest";
 import { bundledLanguagesInfo } from "shiki/langs";
 import { themeNames } from "@shikijs/themes";
-import { codeTextLanguageNames, codeTextThemeNames } from "./code-text-options";
+import { codeTextLanguageNames, codeTextThemeNames } from "./options";
 
 test("matches Shiki's bundled language and theme IDs", () => {
   expect(codeTextLanguageNames).toEqual([
