+++ packages/sdk-components-react/src/code-text/template.tsx
@@ -1,0 +1,12 @@
+import { $, type TemplateMeta } from "@webstudio-is/template";
+
+export const CodeText: TemplateMeta = {
+  category: "typography",
+  template: (
+    <$.CodeText
+      code={'const status = "ready";'}
+      language="javascript"
+      theme="github-light"
+    />
+  ),
+};
