+++ packages/sdk-components-react/tsconfig.dts.json
@@ -3,8 +3,8 @@
   "include": [
     "src/index.ts",
     "src/components.ts",
-    "src/code-text.tsx",
-    "src/code-text.build.ts",
+    "src/code-text/index.tsx",
+    "src/code-text/build.ts",
     "src/build-hooks.ts",
     "src/metas.ts",
     "src/hooks.ts",
